Move transformation helper functions and tests to transform.py
